refactor(frontend): split oversized settings surfaces
